[#50] Move argparsing to a subdirectory to be able to extract model classes into a model.py file
